+++ bocadillo/__init__.py
@@ -1,10 +1,7 @@
 from .api import API
-from .http import HTTPError
-from .media import Media
-from .middleware import Middleware
+from .http import *
 from .recipes import Recipe
 from .staticfiles import static
-from .views import view
-from .websockets import WebSocket, WebSocketDisconnect
+from .websocket import WebSocket, WebSocketDisconnect
 
 __version__ = "0.9.0"
