+++ bocadillo/http/views.py
@@ -2,10 +2,10 @@ import inspect
 from functools import partial, wraps
 from typing import List, Union, Any, Awaitable, Callable, Dict
 
-from .compat import call_async, camel_to_snake
-from .constants import ALL_HTTP_METHODS
 from .request import Request
 from .response import Response
+from ..compat import call_async, camel_to_snake
+from ..constants import ALL_HTTP_METHODS
 
 Handler = Callable[[Request, Response, Any], Awaitable[None]]
 MethodsParam = Union[List[str], all]
