+++ tests/test_http_media_handlers.py
@@ -3,7 +3,7 @@ import json
 import pytest
 
 from bocadillo import API, Media
-from bocadillo.media import UnsupportedMediaType
+from bocadillo.http.media import UnsupportedMediaType
 
 
 def test_defaults_to_json(api: API):
